+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Add `exchange_rate_quote_id` parameter to `Transaction#create`
 * Add `exchange_rate_quote_id` attribute to `Transaction`
 * Add `default?` to `PaymentMethodNonce` (thanks @klouvas)
+* Add error code `TaxAmountIsRequiredForAibSwedish` for attribute `tax_amount` in `transaction` key for AIB:Domestic Transactions in Sweden
 
 ## 4.1.0
 * Add `payment_reader_card_details` parameter to `Transaction.sale`